remove casting warnings
git-svn-id: https://svn.aspect.dealii.org/trunk/aspect@662 244a8438-9d6d-46c2-a098-b74c898f6b90

+++ source/adiabatic_conditions.cc
@@ -144,7 +144,7 @@ namespace aspect
   {
     const double z = point_to_depth_converter(p);
 
-    const unsigned int i = static_cast<double>(z/delta_z);
+    const unsigned int i = static_cast<unsigned int>(z/delta_z);
     Assert (i >= 0, ExcInternalError());
     Assert (i+1 < pressures.size(), ExcInternalError());
 
@@ -162,7 +162,7 @@ namespace aspect
   {
     const double z = point_to_depth_converter(p);
 
-    const unsigned int i = z/delta_z;
+    const unsigned int i = static_cast<unsigned int>(z/delta_z);
     Assert (i >= 0, ExcInternalError());
     Assert (i+1 < temperatures.size(), ExcInternalError());
